moved chat files to backend folder
