+++ package.json
@@ -7,10 +7,10 @@
     "test": "test"
   },
   "scripts": {
-    "test": "lab -t 100 -r lcov > lcov.info",
+    "test": "lab -t 100",
     "start": "nodemon server.js",
     "jshint": "jshint -c .jshintrc --exclude-path .jshintignore .",
-    "coverage": "./node_modules/.bin/istanbul cover ./test/*.js"
+    "coverage": "./node_modules/.bin/istanbul cover ./test/*.js && ./node_modules/.bin/istanbul check-coverage --statements 90 --functions 100 --lines 90 --branches 85"
   },
   "pre-commit": [
     "jshint",
