+++ test/chat-test.js
@@ -1,8 +1,9 @@
 var Code = require('code');
 var Lab = require('lab');
 var lab = exports.lab = Lab.script();
-var handlers = require('../handlers.js');
+var handlers = require('../backend/handlers.js');
 var Shot = require('shot');
+var server = require('../server.js');
 
 // some niceties
 var describe = lab.describe;
@@ -24,7 +25,7 @@ lab.test("chat.connected test", function(done) {
   };
   server.inject(options, function(response) {
     Code.expect(response.statusCode).to.equal(200);
-    Code.expect(response.result).to.equal("<!DOCTYPE html><html><head><meta charset='utf-8'><title></title></head><body><p>testuser2</p></body></html>");
+    Code.expect(response.result).to.equal("<!DOCTYPE html><html><head><meta charset=\"utf-8\"><title>chatbox</title></head><body><div id=\"chatbox\"><div id=\"pastchat\"></div><textarea id=\"message\" rows=\"3\" cols=\"40\"></textarea><button onclick=\"postMessage()\">Send</button></div><script src=\"https://cdn.socket.io/socket.io-1.2.0.js\"></script><script src=\"../public/js/script.js\"></script></body></html>");
     done();
   });
 });
